[PAGOPA-2800] fix unit tests
